requirements/common.txt
--extra-index-url https://test.pypi.org/simple/
chatgpt-mohdfareed
python-telegram-bot
python-dotenv
telethon
qrcode
